refactor: implement new aws provider
