+++ provider/v2/aws/scanestimation/pricefetcher_test.go
